+++ Example/Sources/View Controllers/ChatViewController.swift
@@ -24,7 +24,7 @@ SOFTWARE.
 
 import UIKit
 import MessageKit
-import MessageInputBar
+import InputBarAccessoryView
 
 /// A base class for the example controllers
 class ChatViewController: MessagesViewController, MessagesDataSource {
@@ -111,7 +111,7 @@ class ChatViewController: MessagesViewController, MessagesDataSource {
     func configureMessageInputBar() {
         messageInputBar.delegate = self
         messageInputBar.inputTextView.tintColor = .primaryColor
-        messageInputBar.sendButton.tintColor = .primaryColor
+        messageInputBar.sendButton.setTitleColor(.primaryColor, for: .normal)
     }
     
     // MARK: - Helpers
@@ -142,7 +142,7 @@ class ChatViewController: MessagesViewController, MessagesDataSource {
     
     // MARK: - MessagesDataSource
     
-    func currentSender() -> Sender {
+    func currentSender() -> SenderType {
         return SampleData.shared.currentSender
     }
     
@@ -278,23 +278,23 @@ extension ChatViewController: MessageLabelDelegate {
 
 // MARK: - MessageInputBarDelegate
 
-extension ChatViewController: MessageInputBarDelegate {
-    
-    func messageInputBar(_ inputBar: MessageInputBar, didPressSendButtonWith text: String) {
-          
+extension ChatViewController: InputBarAccessoryViewDelegate {
+
+    func inputBar(_ inputBar: InputBarAccessoryView, didPressSendButtonWith text: String) {
+
         for component in inputBar.inputTextView.components {
-            
+
+            let user = SampleData.shared.currentSender
             if let str = component as? String {
-                let message = MockMessage(text: str, sender: currentSender(), messageId: UUID().uuidString, date: Date())
+                let message = MockMessage(text: str, user: user, messageId: UUID().uuidString, date: Date())
                 insertMessage(message)
             } else if let img = component as? UIImage {
-                let message = MockMessage(image: img, sender: currentSender(), messageId: UUID().uuidString, date: Date())
+                let message = MockMessage(image: img, user: user, messageId: UUID().uuidString, date: Date())
                 insertMessage(message)
             }
-            
+
         }
         inputBar.inputTextView.text = String()
         messagesCollectionView.scrollToBottom(animated: true)
     }
-
 }